Mejoras en endpoints por rol y de rendimiento de apis
Añadir un sistema de notificaciones y refactorizar la respuesta de registro de usuarios

- Se implementó la funcionalidad de creación y recuperación de notificaciones.
- Se añadieron las entidades Notificación y TipoNotificación con sus correspondientes DTO e interfaces de repositorio.
- Se actualizó AuthController para devolver los datos del usuario al registrarse.
- Se refactorizaron varios métodos del repositorio para usar enteros en lugar de largos para los ID.
- Se mejoró EmprendimientoServiceImpl para notificar a los administradores sobre solicitudes nuevas o actualizadas.

+++ src/main/java/com/example/eureka/entrepreneurship/dto/VistaEmprendedorDTO.java
@@ -18,5 +18,5 @@ public class VistaEmprendedorDTO {
     private String observaciones;
     private String motivoRechazo;
     private Boolean tieneSolicitudActiva;
-    private Long solicitudId;
+    private Integer solicitudId;
 }
